Use a singleton for DynamoDbMasterMonitor
We encountered an issue where the MasterMonitor's `shutdown` method was
called somewhere.  This led to clients relying on updates to the
MasterMonitor stopping receiving updates but they had no mechanism of
knowing that the monitor they were subscribed to was no longer updating
because the observable was not completed and getting the latest version
would continue to return the last known value (even though that value
was not updated).

We were deciding between either using a Singleton or tracking down
`shutdown` calls to `MasterMonitor` and ensuring that usages could
resubscribe if necessary.  The latter seemed like it would be prone to
breakage in the future (e.g. particularly with the master observable).
Additionally, a Singleton pattern seemed fitting given that almost every
reasonable use case of Mantis relies on a sustained connection to the
Mantis Master.  So there's no reason to expect that shutting down the
executor pool for updating the master is necessary.  Additionally, using
a Singleton avoids any unbounded resource leakage (e.g. it is still safe
to create N MasterMonitors).
